# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Idealist project API Docs.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "edgardanielgd123@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/team": {
            "post": {
                "description": "Create a new team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Create a new team",
                "parameters": [
                    {
                        "description": "Team Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crud.TeamCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/team/members": {
            "get": {
                "description": "Get all team members",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Get all team members",
                "parameters": [
                    {
                        "description": "Team data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crud.TeamListMembers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a member to a team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Add a member to a team",
                "parameters": [
                    {
                        "description": "New member data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crud.TeamAddMember"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a member from a team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Team"
                ],
                "summary": "Create a new team",
                "parameters": [
                    {
                        "description": "Old member data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crud.TeamRemoveMember"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/user": {
            "post": {
                "description": "Create a new user given username and email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crud.UserCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/user/anonymous": {
            "post": {
                "description": "Create a new user given username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "Anonymous User Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/crud.AnonymousUserCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "crud.AnonymousUserCreate": {
            "type": "object",
            "required": [
                "temp_name"
            ],
            "properties": {
                "temp_name": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "crud.TeamAddMember": {
            "type": "object",
            "required": [
                "member_id",
                "team_id",
                "user_id"
            ],
            "properties": {
                "member_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "team_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "user_id": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "crud.TeamCreate": {
            "type": "object",
            "required": [
                "team_name",
                "user_id"
            ],
            "properties": {
                "team_name": {
                    "type": "string",
                    "maxLength": 255
                },
                "user_id": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "crud.TeamListMembers": {
            "type": "object",
            "required": [
                "team_id",
                "user_id"
            ],
            "properties": {
                "team_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "user_id": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "crud.TeamRemoveMember": {
            "type": "object",
            "required": [
                "member_id",
                "team_id",
                "user_id"
            ],
            "properties": {
                "member_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "team_id": {
                    "type": "string",
                    "maxLength": 255
                },
                "user_id": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "crud.UserCreate": {
            "type": "object",
            "required": [
                "email",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "username": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}